Don't check url certs and suppress warnings.
We set `verify=False` when we pull thumbnail descriptions and image URLs
using requests in order to prevent failure due to invalid SSL
certificates.
This fix also supresses noisy Warnings.
Also, add a comment to alert the user that we are not checking certs,
in the event they are using another data set that is not secure.
Also, fix syntax issue from previous patch (too many quotation
marks).

Closes: #42